refactor: move editoractions into util directory
