+++ core/src/main/java/dk/gtz/graphedit/view/EditorSettingsController.java
@@ -5,7 +5,7 @@ import org.slf4j.LoggerFactory;
 
 import atlantafx.base.controls.Tile;
 import atlantafx.base.controls.ToggleSwitch;
-import dk.gtz.graphedit.tool.EditorActions;
+import dk.gtz.graphedit.util.EditorActions;
 import dk.gtz.graphedit.util.InspectorUtils;
 import dk.gtz.graphedit.viewmodel.ViewModelEditorSettings;
 import dk.yalibs.yadi.DI;
