+++ core/src/main/java/dk/gtz/graphedit/view/ProjectSettingsController.java
@@ -7,7 +7,7 @@ import org.slf4j.LoggerFactory;
 
 import atlantafx.base.controls.Tile;
 import atlantafx.base.theme.Styles;
-import dk.gtz.graphedit.tool.EditorActions;
+import dk.gtz.graphedit.util.EditorActions;
 import dk.gtz.graphedit.util.InspectorUtils;
 import dk.gtz.graphedit.viewmodel.ViewModelProject;
 import dk.yalibs.yadi.DI;
