+++ core/src/main/java/dk/gtz/graphedit/view/RunTargetsEditorController.java
@@ -12,7 +12,7 @@ import org.slf4j.LoggerFactory;
 import atlantafx.base.controls.Tile;
 import atlantafx.base.controls.ToggleSwitch;
 import atlantafx.base.theme.Styles;
-import dk.gtz.graphedit.tool.EditorActions;
+import dk.gtz.graphedit.util.EditorActions;
 import dk.gtz.graphedit.util.InspectorUtils;
 import dk.gtz.graphedit.viewmodel.ViewModelEnvironmentVariable;
 import dk.gtz.graphedit.viewmodel.ViewModelProject;
